genx_platform/genx_components/microservices/llm/requirements.lock.txt
# Generate this file with:
# pip freeze > requirements.lock.txt

# Core dependencies with exact versions
grpcio==1.60.0
grpcio-tools==1.60.0
grpcio-health-checking==1.60.0
protobuf==4.25.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ML/AI with exact versions
transformers==4.36.2
torch==2.1.2
tokenizers==0.15.0
accelerate==0.25.0
sentencepiece==0.1.99
safetensors==0.4.1
huggingface-hub==0.20.1

# OpenTelemetry with exact versions
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-grpc==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-prometheus==0.43b0

# Utilities with exact versions
psutil==5.9.7
prometheus-client==0.19.0
numpy==1.24.4
tqdm==4.66.1